% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static_map.R
\name{get_static_map}
\alias{get_static_map}
\title{Get static map}
\usage{
get_static_map(
  area,
  map_style = "mapbox://styles/mapbox/dark-v10",
  mapbox_api_access_token = Sys.getenv("MAPBOX_ACCESS_TOKEN"),
  retina = TRUE,
  scale_ratio = 1
)
}
\arguments{
\item{scale_ratio}{}
}
\description{
Get static map
}
