% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{is_zero}
\alias{is_zero}
\title{is_zero}
\usage{
is_zero(x)
}
\arguments{
\item{x}{the number to compare to 0}
}
\value{
{boolean}
}
\description{
using all.equal to allow for areas that are very very small to be considered
as equal to zero.
}
