% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer.R
\name{layer_mapbox}
\alias{layer_mapbox}
\title{Layer mapbox}
\usage{
layer_mapbox(
  area,
  map_style = mapbox_dark(),
  mapbox_api_access_token = Sys.getenv("MAPBOX_ACCESS_TOKEN"),
  retina = TRUE,
  scale_ratio = 1,
  area_buffer = 0,
  purge_cache = FALSE
)
}
\arguments{
\item{area}{the map area}

\item{map_style}{mapbox style, see \code{stylebox::mapbox_styles}}

\item{mapbox_api_access_token}{mapbox api access token}

\item{retina}{render the map at 2x scale if \code{TRUE}}

\item{scale_ratio}{ratio to scale the output image; \code{scale_ratio = 1} will give the largest image and smallest zoom.}

\item{area_buffer}{a buffer to appear around the \code{area} geometry in meters (web mercator projection, EPSG 3857).
Use this to create space around your data in the map visual.}

\item{purge_cache}{forget cached api calls and responses before making api call if TRUE.}
}
\description{
Create a mapbox map ggplot layer
}
