% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_styles.R
\docType{data}
\name{map_styles}
\alias{map_styles}
\alias{mapbox_streets}
\alias{mapbox_outdoors}
\alias{mapbox_light}
\alias{mapbox_dark}
\alias{mapbox_satellite}
\alias{mapbox_satellite_streets}
\title{Map styles}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
mapbox_streets

mapbox_outdoors

mapbox_light

mapbox_dark

mapbox_satellite

mapbox_satellite_streets
}
\description{
Map styles
}
\seealso{
\url{https://docs.mapbox.com/api/maps/#mapbox-styles}
}
\keyword{datasets}
