% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static_map.R
\name{get_static_map}
\alias{get_static_map}
\title{Get static Mapbox map as a georeferenced raster}
\usage{
get_static_map(
  area,
  map_style = mapbox_dark(),
  mapbox_api_access_token = Sys.getenv("MAPBOX_ACCESS_TOKEN"),
  retina = TRUE,
  scale_ratio = 1,
  area_buffer = 0,
  purge_cache = FALSE
)
}
\arguments{
\item{area}{The map area}

\item{map_style}{A Mapbox style, see \code{stylebox::mapbox_styles}}

\item{mapbox_api_access_token}{Mapbox api access token}

\item{retina}{Render the map at 2x scale}

\item{scale_ratio}{Ratio to scale the output image; \code{scale_ratio = 1} will give the largest image and smallest zoom.}

\item{area_buffer}{a buffer to appear around the \code{area} geometry in meters (web mercator projection, EPSG 3857). Use this to create space around your data in the map visual.}

\item{purge_cache}{forget cached api calls and responses before making api call if TRUE.}
}
\description{
Get static Mapbox map as a georeferenced raster
}
