% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static_map.R
\name{get_static_map}
\alias{get_static_map}
\title{Get static map}
\usage{
get_static_map(
  area,
  map_style = mapbox_dark(),
  mapbox_api_access_token = Sys.getenv("MAPBOX_ACCESS_TOKEN"),
  retina = TRUE,
  scale_ratio = 1,
  area_buffer = 0,
  mapbox_logo = TRUE,
  attribution = TRUE,
  purge_cache = FALSE
)
}
\arguments{
\item{area}{the map area}

\item{map_style}{mapbox style, see \code{stylebox::mapbox_styles}}

\item{mapbox_api_access_token}{mapbox api access token}

\item{retina}{render the map at 2x scale if \code{TRUE}}

\item{scale_ratio}{ratio to scale the output image; \code{scale_ratio = 1} will give the largest image and smallest zoom.}

\item{area_buffer}{a buffer to appear around the \code{area} geometry in meters (web mercator projection, EPSG 3857).
Use this to create space around your data in the map visual.}

\item{mapbox_logo}{include mapbox logo on the image if TRUE}

\item{attribution}{include attribution on the image if TRUE.
You still have a legal responsibility to attribute maps that use OpenStreetMap data, which includes most maps from Mapbox.
If you specify attribution=FALSE, you are legally required to include proper attribution elsewhere on the webpage or document.
\url{https://docs.mapbox.com/api/maps/#static-images} and \url{https://docs.mapbox.com/help/how-mapbox-works/attribution/#static--print}}

\item{purge_cache}{forget cached api calls and responses before making api call if TRUE.}
}
\description{
Get static Mapbox map as a georeferenced {stars} raster
}
